refactor: Restructure API service architecture
- Move service.ts to api/service.ts for better organization
- Separate SWR hooks into api/hooks.ts for modular architecture
- Split core API client from SWR functionality
- Update all import paths from @/service to @/api/hooks
- Maintain clean separation of concerns:
  - api/service.ts: Core API client and types
  - api/hooks.ts: SWR-based React hooks
- All components updated with new import paths
- Build successful with no compilation errors
